requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-ckeditor==5.4.0
django-js-asset==1.1.0
geoip2==2.8.0
idna==2.6
maxminddb==1.3.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
requests==2.18.4
urllib3==1.22
uWSGI==2.0.17

